backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.9.0
pydantic-core==2.23.0
